requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Django==3.2.25
djangorestframework==3.15.1
